+++ src/input-number/test/input-number.spec.ts
@@ -389,6 +389,13 @@ describe('input-number component', () => {
         fixture.detectChanges();
 
         expect(downHandlerElement.classList.contains('disabled')).toBe(false);
+
+        inputNumberComponentInstance.thyMin = null;
+        fixture.detectChanges();
+        flush();
+        fixture.detectChanges();
+
+        expect(downHandlerElement.classList.contains('disabled')).toBe(false);
     }));
 
     it('should remove disabled in up handler after switch thyMax', fakeAsync(() => {
@@ -406,6 +413,12 @@ describe('input-number component', () => {
         flush();
         fixture.detectChanges();
 
+        expect(upHandlerElement.classList.contains('disabled')).toBe(false);
+
+        inputNumberComponentInstance.thyMax = null;
+        fixture.detectChanges();
+        flush();
+        fixture.detectChanges();
         expect(upHandlerElement.classList.contains('disabled')).toBe(false);
     }));
 });