+++ src/input/input.component.ts
@@ -88,14 +88,6 @@ export class ThyInputComponent implements ControlValueAccessor, OnInit {
     @Input() readonly = false;
 
     /**
-     *  输入字段是否应该启用自动完成功能
-     */
-    @Input()
-    set thyAutocomplete(value: boolean) {
-        this.autocomplete = value;
-    }
-
-    /**
      * focus 聚焦事件
      */
     @Output() focus: EventEmitter<Event> = new EventEmitter<Event>();
@@ -120,8 +112,6 @@ export class ThyInputComponent implements ControlValueAccessor, OnInit {
     public _type = 'text';
 
     public value: string;
-
-    public autocomplete: boolean;
 
     public showLabel: boolean;