+++ src/message/message-queue.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
+import { ThyAbstractMessageQueue } from './abstract';
 import { ThyMessageRef } from './message-ref';
-import { ThyGlobalMessageConfig, ThyMessageConfig, THY_MESSAGE_DEFAULT_CONFIG } from './message.config';
+import { ThyGlobalMessageConfig, THY_MESSAGE_DEFAULT_CONFIG, THY_MESSAGE_DEFAULT_CONFIG_VALUE } from './message.config';
 
 /**
  * @internal
@@ -9,36 +9,11 @@ import { ThyGlobalMessageConfig, ThyMessageConfig, THY_MESSAGE_DEFAULT_CONFIG } 
 @Injectable({
     providedIn: 'root'
 })
-export class ThyMessageQueue {
-    queues$ = new BehaviorSubject<ThyMessageRef[]>([]);
-
-    get queues() {
-        return this.queues$.getValue();
-    }
-
-    constructor(@Inject(THY_MESSAGE_DEFAULT_CONFIG) private defaultConfig: ThyGlobalMessageConfig) {
-        this.defaultConfig = defaultConfig;
-        this.queues$ = new BehaviorSubject([]);
-    }
-
-    add(messageRef: ThyMessageRef) {
-        const queues = this.queues$.getValue();
-        if (this.queues.length >= this.defaultConfig.maxStack) {
-            const closedRef = queues.shift();
-            closedRef.close();
-        }
-        this.queues$.next([...queues, messageRef]);
-    }
-
-    remove(id: string) {
-        if (!id) {
-            this.queues.forEach(item => item.close());
-            this.queues$.next([]);
-        } else {
-            const removeItem = this.queues.find(item => item.id === id);
-            removeItem?.close();
-            const afterRemovedQueues = this.queues.filter(item => item.id !== id);
-            this.queues$.next(afterRemovedQueues);
-        }
+export class ThyMessageQueue extends ThyAbstractMessageQueue<ThyMessageRef> {
+    constructor(@Inject(THY_MESSAGE_DEFAULT_CONFIG) defaultConfig: ThyGlobalMessageConfig) {
+        super({
+            ...THY_MESSAGE_DEFAULT_CONFIG_VALUE,
+            ...defaultConfig
+        });
     }
 }