+++ src/message/message.component.ts
@@ -1,11 +1,8 @@
-import { Component, Input, HostBinding, OnInit, HostListener, OnDestroy, NgZone } from '@angular/core';
+import { Component, Input, HostBinding, NgZone } from '@angular/core';
 import { trigger, state, style, animate, transition } from '@angular/animations';
 import { ThyMessageConfig } from './message.config';
 import { ThyMessageQueue } from './message-queue.service';
-
-const ANIMATION_IN_DURATION = 100;
-const ANIMATION_OUT_DURATION = 150;
-const HIDE_STYLE = { transform: 'translateX(0)', opacity: 0, height: 0, paddingTop: 0, paddingBottom: 0, margin: 0 };
+import { ANIMATION_IN_DURATION, ANIMATION_OUT_DURATION, HIDE_STYLE, ThyAbstractMessageComponent } from './abstract';
 
 /**
  * @internal
@@ -24,77 +21,22 @@ const HIDE_STYLE = { transform: 'translateX(0)', opacity: 0, height: 0, paddingT
                 animate(ANIMATION_IN_DURATION)
             ]),
             transition('flyIn => componentHide', [animate(ANIMATION_OUT_DURATION, style(HIDE_STYLE))]),
-            transition('flyIn => void', [animate(ANIMATION_IN_DURATION, style(HIDE_STYLE))]),
 
             state('componentHide', style(HIDE_STYLE))
         ])
     ]
 })
-export class ThyMessageComponent implements OnInit, OnDestroy {
-    @HostBinding('@flyInOut') flyInOut = 'flyIn';
+export class ThyMessageComponent extends ThyAbstractMessageComponent<ThyMessageConfig> {
+    @HostBinding('@flyInOut') animationState = 'flyIn';
 
     config: ThyMessageConfig;
-
-    iconName = '';
-
-    private closeTimer: any;
 
     @Input()
     set thyConfig(value: ThyMessageConfig) {
         this.config = value;
     }
 
-    constructor(private _ngZone: NgZone, private messageQueue: ThyMessageQueue) {}
-
-    ngOnInit() {
-        const iconName = {
-            success: 'check-circle-fill',
-            info: 'info-circle-fill',
-            warning: 'waring-fill',
-            error: 'close-circle-fill'
-        };
-
-        this.iconName = iconName[this.config.type];
-        this.createCloseTimer();
-    }
-
-    @HostListener('mouseenter')
-    mouseenter() {
-        if (this.config.pauseOnHover) {
-            this.clearCloseTimer();
-        }
-    }
-
-    @HostListener('mouseleave')
-    mouseleave() {
-        if (this.config.pauseOnHover) {
-            this.createCloseTimer();
-        }
-    }
-
-    close() {
-        this._ngZone.runOutsideAngular(() => {
-            this.flyInOut = 'componentHide';
-            setTimeout(() => {
-                this.messageQueue.remove(this.config.id);
-            }, ANIMATION_OUT_DURATION);
-        });
-    }
-
-    private createCloseTimer() {
-        if (this.config.duration) {
-            this.closeTimer = setInterval(() => {
-                this.clearCloseTimer();
-                this.close();
-            }, this.config.duration);
-        }
-    }
-
-    private clearCloseTimer() {
-        clearInterval(this.closeTimer);
-    }
-
-    ngOnDestroy() {
-        this.clearCloseTimer();
+    constructor(ngZone: NgZone, messageQueue: ThyMessageQueue) {
+        super(ngZone, messageQueue);
     }
 }