+++ src/notify/notify.options.ts
@@ -1,7 +1,0 @@
-import { ThyAbstractOverlayOptions } from 'ngx-tethys/core';
-
-export const notifyAbstractOverlayOptions: ThyAbstractOverlayOptions = {
-    name: 'notify',
-    animationEnabled: true,
-    disposeWhenClose: true
-};